+++ components/Checkout/CustomDonationInput.tsx
@@ -1,6 +1,5 @@
 import { ChangeEvent } from 'react';
 import { TextField } from '@react-md/form';
-import { Typography } from '@react-md/typography';
 import { AttachMoneySVGIcon } from '@react-md/material-icons';
 
 type CustomDonationInputProps = {
@@ -23,12 +22,9 @@ export const CustomDonationInput = ({
 }: CustomDonationInputProps) => {
   return (
     <label>
-      <Typography type="headline-5">
-        Specify the amount you would like to donate
-      </Typography>
       <TextField
         id="price"
-        theme="underline"
+        theme="outline"
         leftChildren={<AttachMoneySVGIcon />}
         className={className}
         type="number"
@@ -37,6 +33,7 @@ export const CustomDonationInput = ({
         min={min}
         max={max}
         onChange={onChange}
+        label="Amount"
       />
     </label>
   );